입력창 스택화, 직접입력 창에서 버튼을 누르면 NumberCheck.js로 넘어가기 성공
