+++ package.json
@@ -11,6 +11,7 @@
 		"@react-navigation/bottom-tabs": "^6.0.7",
 		"@react-navigation/material-top-tabs": "^6.0.5",
 		"@react-navigation/native": "^6.0.4",
+		"@react-navigation/native-stack": "^6.2.4",
 		"@react-navigation/stack": "^6.0.11",
 		"eslint-config-prettier": "^8.3.0",
 		"eslint-plugin-prettier": "^4.0.0",
